Move RITM; Update worker
